requirements.txt
flask
opencv-python-headless
torch
torchvision
Pillow
supervision
